move schemas and make utility functions.
